+++ src/index.ts
@@ -1,4 +1,4 @@
-export * from './auto-reset-ref'
+export * from './use-reset-ref'
 export * from './use-fetch-data'
 export * from './use-lock-screen'
 export * from './use-mutation'